Function rust_experiments::cities::gen_cities_uset

pub fn gen_cities_uset(size: usize, max_roads_per_distance: usize) -> Vec<usize>

Generates a city map.

Creates more "snaky" maps than gen_cities. The computed distances vectors tend to be long and with low values at each index. Slower, and more complex. Uses USet.

Arguments

size - the total number of cities

max_roads_per_distance - the maximum number of roads leading to all cities with the same distance to the capital

Example

use rust_experiments::cities::*;
let city_vec = gen_cities_uset(10, 3);
assert_eq!(city_vec.len(), 10);